Update patch set 1
Patch Set 1:

This looks like it will prevent the failing build.  The core reason that the build was failing is that the paths have changed. The code was supposed do go into an existing path.

I will take a look at the script with Larry and see if we can address the incorrect path. Not just the creation of the path.

Patch-set: 1