Example Path update ticket#39430
The examples are no longer shipped as android/examples.
They are now java/examples-android

Replace all references to
 android/examples
in
 jjb/iotivity/iotivity-upnp-jobs.yaml
with
 java/examples-android

Remove all references to
 android/example
from
 jjb/iotivity/include-raw-iotivity-bootstrap.sh

My first attempt to codify the prose representation of the concept of
s/$search/$replace/g for some value of $search = /android/i and some
value of $replace = /java/i as related to directory structure.

Change-Id: I217e4fafcdf43c7a74f9b1114a17f05fd6860686
Signed-off-by: C.J. Collier <cjcollier@linuxfoundation.org>