Update patch set 2
Patch Set 2:

> (1 comment)
 > 
 > I know this has been merged, but my comment here is for future
 > improvement.

I'd be surprised that the list from run.bat is complete. It seems reasonable to me to start with those tests that we have seen deadlocking so far. But, I agree that we might need to extend this list in the future.

Patch-set: 2
Reviewer: Gerrit User 7511 <7511@c6509b8a-a471-4ec1-8e5f-9aed2654acb5>
Label: Verified=0