Fix Swarm Connections and Save Gerrit Hostkey
There is a bug in the swarm plugin that causes unicast requests to try
and upgrade to TLS. This fails obviously with the setup because builders
are not connecting over the internet. Disabling the ssl verification
keeps the NullPointerException from happening

Pre-emptively grabbing the Gerrit ssh-hostkey should save us a headache
with ssh-agent.

Change-Id: Ifb34686a6485263fd172358938787e226f560110
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>