(syntax error fix) Terminate tests in a loop before and after test run
When terminating tests, use a single instance of the command to
execute.  This reduction of duplicate code reduces the likelihood of
human error when changing code in one instance which should be applied
to all instances.  This also results in fewer lines of code.

Run the test process cleanup script before the build in addition to
after.

Change-Id: If378582ed95aaa605b4944ef9937347df8498c1d
Signed-off-by: C.J. Collier <cjcollier@linuxfoundation.org>