Bump tinycbor to 0.5.0 for master
The various places found will now download all the known
versions and then make a link to a version chosen based om
git branch.

This windows file: jjb/iotivity/include-raw-iotivity-windows-bootstrap.bat
was not updated, it would seem to require some conditional code to
match the .sh files, and I'm not comfortable trying that out.
Will add to the patch if anyone has suggestions.

Note there was discussion elsewhere adding a Python script to
"fix" the right version of download/unpack bundles, but this
patch is not proposing that - handled separately.

Change-Id: I89e4f57c06f3516e4a4d738e7435cf6278bffe5b
Signed-off-by: Mats Wichmann <mats@linux.com>