Remove unnecessary clean-up
After the migration to single-use slaves, we no longer need to clean
out the unit test processes.  Rather than attempting to return the
build instance to a state where it can be re-used to run jobs, we now
only use the builder for a single job and then throw it away.  No need
to clean up extra processes now.

Change-Id: Id0b5c8263a07e5af0888d1b67e3a99ce464141b1
Signed-off-by: C.J. Collier <cjcollier@linuxfoundation.org>