pdate Packer Configuration
Updates common-packer to the latests version, bumps global-jjb to
v0.20.0 so the packer-version can be configured, and removed outdated
packer scripts.

Set the cloud_auth_url to come from var files, and set the domain for
Openstack builds to 'Default'

Removes the requirement on chaining of packer builds and need to update
json files after new images are created.

Change-Id: Ice82ce25d5823018dbef0e6562f452456d06b46c
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>