Update patch set 21
Patch Set 21:

Hi Suresh, this look like it's good to go, and I saw the build pass in the sandbox.

Next steps: Please split out the change to 'jjb/iotivity/include-raw-iotivity-bootstrap.sh' to a separate patch. We can't have that change merged in at the same time as the image it relies on won't have been built yet.

There is still a body missing from the commit message. Please add at least a sentence describing why this change is being made.

Once the bootstrap script is split out, I'll remove my -2 and this patch should be good to go.

Patch-set: 21