Update patch set 5
Patch Set 5: Code-Review-1

(2 comments)

The default install should be for vs2015 (Or is it 2017 now?) and be part of the windows.yaml tasks (the variable stored in defaults.yaml).

For other versions we can override the variable (vs_version) to the version of VS we may need.

Patch-set: 5
Label: Code-Review=-1