Update patch set 1
Patch Set 1:

> > Build Successful
 > >
 > > https://jenkins.iotivity.org/job/ci-management-tox-verify-master/335/
 > > : SUCCESS
 > >
 > > https://jenkins.iotivity.org/job/ci-management-jjb-verify/309/ :
 > > SUCCESS

You don't need to, nor should you, +1 your own patches. It's implied by you uploading the change that you approve it. :)

You can -2 if what you have is still a work-in-progress, and remove it when your patch is ready for review, but that's only if you feel it's not obvious from your change more testing needs to be done or that the patch isn't complete.

Patch-set: 1