Update patch set 8
Patch Set 8:

I believe there is a update to global-jjb that will add a verified+1 for views, so we'll have to wait for global-jjb to be updated first. Or it could be the updated global-jjb verifies any *.yaml changes which is why this isn't being picked up yet.

Patch-set: 8
Reviewer: Gerrit User 1154 <1154@c6509b8a-a471-4ec1-8e5f-9aed2654acb5>
Label: Verified=0