Update patch set 2
Patch Set 2: Code-Review-1

The commit message should be changed to state this is an upgrade to global-jjb, not a preparation of the repo.

Before upgrading to 0.28.0, we should upgrade JJB. Views weren't introduced till 2.0.5 and IoTivity currently is using 2.0.3. So the verification you're getting is actually a false positive (there is nothing actually checking the view code is correct)

Patch-set: 2